Implement some spawn systems
This introduces 3 new spawn systems, that can be voted with `ag_spawn_system`,
which is followed by a number that corresponds to the spawn system, so:
* ag_spawn_system 0 - votes the Classic spawn system (it's the default one)
* ag_spawn_system 1 - votes the Random spawn system
* ag_spawn_system 2 - votes the Far spawn system
* ag_spawn_system 3 - votes the Position-aware spawn system (abbreviated PA)

The Random spawn system is self-explaining, players spawn in any spawnpoint,
it can even be the same where someone has just spawned.

The Far spawn system chooses randomly some far spawnpoint, avoiding recently
used ones.

The Position-aware spawn system classifies spawnpoints into different lists
and randomly picks one list, with different probabilities that can be voted.
Then it chooses one random spot from that list, avoiding recently used ones.
These lists are:
* Visible spots: spots that can be seen (or almost) by some enemy
* Audible spots: spots where some enemy can hear someone spawn
* Barely-audible spots: this is just a fallback list, it's mostly safe spots
* Safe spots: these are just far spots that are not visible nor audible

Note: you can spawn on top of another player, they won't get telefragged
or anything. We can change this later.


It's important to note that the default spawn system will be the Classic one,
the spawn system that we've always used in AG. The rest can be voted, and
the spawn system changes inmediately after the vote being accepted.

Some spawn systems do not fit some maps, gamemodes or formats, or combinations
of these; but they might be better than the Classic one in some cases.
The Classic one is excellent for competitive matches, it should be the one
to be used in tournaments, IMO, because we're used to it and our current
strategies depend on it, it has a lot of weight in our positioning and
decision-making in crossfire for example. So these new spawn systems
should NOT be used in tournaments, at least for a very long time that we
need to test where they fit. They will need a lot of tuning and fixing.