Lower default intermission/chat times + Add cvar
Added the `mp_intermission_time` cvar, which tells how many seconds the
whole intermission can take to finish automatically.

So the intermission is divided in 2 parts, and it always has been, but before
this commit only the first part was customizable, and in AG 6.6 not even
the first part. That first part is the chat time, configurable by the
`mp_chattime` cvar, which was set to 10 seconds automatically in AG 6.6
right at the start of the intermission, so whatever value you had configured
for it, it was ignored completely. That was fixed some time ago in AG 6.7,
but the second part (the rest of the intermission) was still hardcoded.

So the first part, the chat time, is where players usually say _gg_, etc.,
so they need a few seconds to chat a bit before the map changes. This part
cannot be skipped. It's now set by default to 6 seconds, as it seems it
was originally thought out by VALVe, and seems kinda reasonable, but maybe
you want to tweak it a bit. Whatever you set, it will be clamped between
1 and 120 seconds, or by the new `mp_intermission_time` cvar, whichever
is more strict; so those are the limits for `mp_chattime`.

The second part is just the rest of the intermission. After the chat time
is over, any player that tries to attack, jump, duck or use, will put an
end to the intermission and the server will load the next map, so just
one player in the server has to do this. The problem is that sometimes,
no player does this, specially in bhop/kz servers, where some unexperienced
player might just tab out of the game and wait for the game to load the next
map. They don't know how to skip the second part and they just had to wait
2 minutes for it to end, which is boring and probably make them think the
game is bad, etc. Those 2 minutes were hardcoded, so now it's configurable
via the `mp_intermission_time`, which will be clamped between 1 and 120
seconds. By default it's now set to 15 seconds (it was 120 before).

If both `mp_chattime` and `mp_intermission_time` have the same
value, then all the intermission will be just chat time, unskippable, and
it will change map automatically when it's over. You have to consider that
you can still chat during the second part of the intermission, after the
chat time. It just won't change unless you attack/jump/duck/use. So some
players might want a bit more time to chat, maybe in teamplay it won't
be useful as some player will surely skip the intermission, but in 1v1
if both players have stuff to say they won't want to skip inmediately,
they might chat for a few seconds more, that's why the default chat time
is now 6 seconds but the whole intermission can take up to 15 seconds, so
9 extra seconds to chat for those cases. Maybe it's too low, we'll see.