Remove questionable command
The command in question can be typed by a player on their client console
while they're in an online server, and makes the server return info
about the kernel and OS versions, which doesn't seem right.

It's not info that a player would need, maybe an admin (AgAdmin) and
still doubtfully. And it can be potentially used by people with bad intentions,
to exploit some vulnerability or use the info in some other malicious way.