Add cvar to auto skip intermission
Server cvar, cannot be voted:
* mp_intermission_skip_auto <0|1|2> - Default: 1

When enabled, it skips either the whole intermission or just the end of it,
only if there's 1 player or none, excluding bots, but accounting for
spectators and proxies. So if it's enabled and there's 1 player and
1 bot, it will be skipped. If there's 1 player and 1 spectator, it
won't be skipped. This is for the case where a player joins a server and
changes map while waiting for his mate(s) to join, so it changes the map
very quick. Also when you're playing in a bhop server alone, and there's
no point in waiting for the intermission, because the score is almost
meaningless in that gamemode and there's no one to chat with at the end.

If the score is meaningful, setting it to 2 might be bad; 1 would be better,
so that in case there's a player practicing against a bot, the chattime is
not skipped and they can see the score during that time, but the rest of
the intermission is skipped automatically.

When it's set to 0 it's disabled, so nothing changes.

When it's set to 1, only the part after the chattime is skipped (the 2nd part
of the intermission).

When it's set to 2, the whole intermission is skipped, including the chattime.