Add votes for extending timelimit
New vote commands:
* agmaxtime - Extends the time limit to the maximum allowed
* agmoretime - Extends the time limit by a certain amount of minutes, defined by server

New cvar:
* sv_ag_vote_extra_timelimit <0|number> - How much time the `agmoretime` vote will add, default 30 minutes. 0 to disable

Quality-of-life votes. Tired of not knowing how much timelimit you can vote for?
Afraid of getting the map changed because you might vote a lower mp_timelimit
than the current one? These new votes come to resolve these issues.

`agmoretime` can actually make the timelimit go beyond the maximum allowed,
that's if it's allowed by `sv_ag_vote_extra_timelimit`. I think it's good
to let the players extend it as much as they want, because otherwise they'll
just reset the map to reset/overcome the limitation, so it's just dumb
putting obstacles to voting the timelimit.

`agmoretime` is designed to extend time by a limited amount, because the
server owner might not want to have the server 24/7 with that map, in the
end they set `sv_ag_vote_mp_fraglimit_high` for a reason. Having to actively
extend the time every X minutes is not fun, but at least you can go over
the max limit and continue playing, while probably still complying with
admin's decision on the max timelimit, because they probably don't care
if the map is still going on as long as people is actively playing on it;
the problem is when it's left empty with that map for a long time, they
probably want to have it changing so people see another map and join the
server to play it, or whatever reason.

`agmaxtime` will vote for the max mp_timelimit or will remove the timelimit
altogether if it's allowed to. Classic case where you want to grind a map
and you don't know how much time you'll spend on it. With this you don't
have to learn every servers' max timelimit anymore.