Add damage and spawn messages for bots
Requires clientside changes. The client has to read a byte at the end
of the TextMsg message, after reading the strings. This byte tells what
type of message it is, so it should have cvars to control whether to
ignore these messages or not, and the damage and spawn types should be
ignored by default, but more might be added in the future, e.g. to be
able to ignore name and model change messages, or join/leave messages.