Implement static bots that respawn upon death
New vote:
* spawnbot - Spawns a bot that doesn't move and respawns when you kill it

New cvars:
* sv_ag_vote_bot <0/1> - Default: 1. Allow adding bots with the spawnbot vote
* sv_ag_bot_limit <number> - Default: 5. Limit the amount of AG bots in a map

This is the AG implementation of MT Bots. A chat message will pop up when
you kill a bot, telling where it has spawned, and another message every
time you damage it. This will be sent to everyone in the server, which is not
ideal, but it's a bit of a hassle to have say commands or similar for
every type of message and stuff, and I don't think these messages should
depend on a serverside cvar, because some players might wanna see while others
won't. So it's up to the clients to implement ignoring capabilities upon
the message types that are now sent on TextMsg

Message types:
-1: type is absent, normal message
0: spawn message
1: damage message