Add flood control for name/model changes
New cvars:
* sv_ag_flood_name_cooldown <number> - Default: 2
* sv_ag_flood_name_spec_cooldown <number> - Default: 60
* sv_ag_flood_model_cooldown <number> - Default: 2
* sv_ag_flood_model_spec_cooldown <number> - Default: 60

These cvars control how many seconds you have to wait to change name/model
again, and there's also specific cvars for when you're on spectator mode,
because some people are very expressive when spectating matches and
they can bother the ones playing the match.

The default 1 minute cooldown for specs is meant to diminish the annoyance,
not to remove spectator expressivity altogether, although some servers may
increase the cooldown to 1200 seconds (20 minutes) to stop them from talking,
but that's up to the admins, I think 1 minute is good for now, we'll see if
it needs tweaking.

The 2 seconds cooldown for players ingame (not spec) is meant to stop
fast name/model changing scripts that have been seen crashing servers
in some ocassions. Again, we'll see if it needs tweaking.