Classify name/team change and game leave messages
The client should now receive the byte that tells if a message is a
name change message, a team change message or a game leave message.

Examples:
* Name change message: `* Player changed name to AAA`
* Team change message: `* Player has changed to team 'blue'`
* Game leave message: `- Player has left the game`

The goal is that clients implement message ignoring capabilities, without
having to check if the text contains certain words. Thus players can easily
toggle if they want to see/hide those messages.

Note: there's no game joining message. That's something that some clients
implemented, it's not something that was originally in the server.