Enforce model or spec for a while
When someone is `agforcespectator`'d or `agforceteam`'d they will be
unable to exit spec or change model respectively until the cooldown
has expired.

Note that the client still thinks they have certain model and not the
forced one, as can be seen when typing `model` or `setinfo`. I haven't
been able to fix that easily without major changes. They just have to
set the model to another thing and back to the desired one if it's bugged
out. The enforcement is cached so that it's still going after reconnecting.

New cvar:
* ag_enforcement_cooldown <number> - Default: 15 (seconds until next spec/model change)