Increase MOTD max length
From 1536 chars to 3000. I experienced overflow issues with over 3000 chars,
that's why it's not 4096 or just unlimited. Clients can receive these long
MOTD without crashing, it will just be capped on their side to 1536

Servers probably should not add crucial info beyond the 1536 characters,
because even if clients get updated with the increased limit, there'll
still be old clients around for some years... and they won't be able to
see the full length of the MOTD. It might still be useful for displaying
a leaderboard or something, where you just print the top 50 and old clients
will see only the top 24, and have to type another command to see further,
like in the HLKZ gamemode