Make respawn time FPS-independent + Fix mp_forcerespawn
New cvar:
* sv_ag_min_respawn_time <0-120> - Default: 0.75 (seconds until respawn)
* sv_ag_forcerespawn_time <number> - Default: 5 (seconds until forced respawn)

Done a few tests about respawn times when it was FPS-dependent, and it
showed inconsistent results, like on average the higher the player FPS,
the faster respawn time, but there are outliers like spawning in 0.68s
at 100 fps instead of 1.2s. The average for 144 fps is 0.833s, and
at times you could spawn in 0.7s, so I made the decision of defaulting
to 0.75s, which is a bit fast but not as much as 200 fps. It's the
equivalent of respawning at 160 fps. I don't want the game to feel slow
so I decided to go for that time.

The minimum respawn time can be as high as 120 seconds (2 minutes), in case
someone sets some crazy value and players can no longer respawn... 120 is
still a lot but someone may create a gamemode where you have to wait more
and more every time you die or level up, like in those MOBA games, so I
figured I would set a max that is somewhat high for such gamemode to give
them margin to do stuff.

Also fixed `mp_respawntime` as it was supposed to be 5 seconds, but it would
be more than that, since it didn't account for the time where you couldn't
respawn yet, which adds up to the 5 seconds making it 6 usually. Now it
should be 5 until player is forced to respawn since the moment they died,
and that's by default, because it's now a cvar that can be tweaked