Match target player by partial name in votes
agallow, agkick, agadmin, agforcespectator and agforceteam now don't require
that you put the whole name of the target player, so it's now more convenient
and a bit easier than issuing `status` to see the #userID and voting by that.
You can also match names that include color codes ("^1", etc.) without
typing them in the vote, you can just ignore them and it should match the
desired player.

If more than 1 player matched your input, the vote won't happen; but you'll
see in console all the matched players with their #userID, so that you can
quickly call the vote again, this time with the #userID which will ensure
that your vote starts successfully with the correct target