Fix forcespec'd players being able to play a match
When a match is going on and you force some player into spectator mode
(with `agforcespectator`), they can get back into the game after the
spectator enforcement time is over, by simply issuing `spectate`. I believe
this was also a thing in v6.6 and it doesn't seem to be right. So now they
should be voted back into the match via `agallow` if they really have to
play it, they will no longer be able to `spectate` back into the game.