CVar overhaul
Now we can keep track of what cvars are there and see if they have changed
from the value that the gamemode, server config, game library or engine
set them initially. This allows us to vote an extensive list of cvars that
affect the gameplay, thus not needing an admin or server owner to make a
new gamemode or something to meet the requirements of a tournament, for
example. Most importantly it allows players to check what gamemode-related
cvars have been changed during a session or map, so that they can make sure
that the server settings meet their needs. They can check a list of cvars
that the gamemode config is using by typing `gamemodevars` in the console,
and a list of cvar changes by typing `settings` in the console.

The votable and tracked cvars at the moment are whitelisted and hardcoded
into the game library. I may allow whitelisting them from a file in the
future, so that server owners can decide what they want to be votable
or not, but I also wish to have this mostly standardized so that players
don't get confused about what is votable and what isn't in each server.

This commit also introduces a bit of trash into the server console if
`developer` is set to 1, but it's only during changelevels and most of
the servers don't need to have `developer 1` (it's actually discouraged).