Implement nuke blocking capabilities
New cvars:
* sv_ag_allow_nuke <0|1> - Default: 1 (don't block nukes)
* sv_ag_nuke_grenade <0|1> - Default: 1 (allow nade nukes)
* sv_ag_nuke_crossbow <0|1> - Default: 1 (allow bolt nukes)
* sv_ag_nuke_rpg <0|1> - Default: 1 (allow rocket nukes)
* sv_ag_nuke_gauss <0|1> - Default: 1 (allow lampgauss and other nukes)
* sv_ag_nuke_egon <0|1> - Default: 1 (allow egon through ceiling)
* sv_ag_nuke_tripmine <0|1> - Default: 1 (allow ghostmines)
* sv_ag_nuke_satchel <0|1> - Default: 1 (allow satchel nukes)
* sv_ag_nuke_snark <0|1> - Default: 1 (allow exploding snark nukes)
* sv_ag_nuke_other <0|1> - Default: 1 (allow any other nuke)
* sv_ag_explosion_fix <0|1> - Default: 0 (don't fix explosion position)

Extended cvars:
* ag_gauss_fix <0|1|2|3> - Default: 0 (allow all selfgauss)


Set `sv_ag_allow_nuke` to 0 to block every nuke, regardless of the specific
values of the other `sv_ag_nuke_*` cvars. When set to 1, it only means that
the specific cvars will be evaluated to decide if the nuke has to be blocked.

Regarding the specific nuke cvars, a few examples:
* `sv_ag_nuke_rpg 0` blocks the nuke produced by launching the rocket over
the ceiling/sky in lost_village2, rockets between boxes (stalkx?), etc.
* `sv_ag_nuke_gauss 0` blocks every lampgauss in crossfire, and should also
block any other kind of nuke you can do with gauss.
* `sv_ag_nuke_tripmine 0` blocks ghostmining, like tripmines in corners or
other spots in datacore.
* `sv_ag_nuke_egon 0` blocks hurting someone on the floor above when you
fire the egon beam at the ceiling, like in crossfire or killbox.

Then we have `sv_ag_explosion_fix`, which when set to 1 tries to position
the explosions produced by nades, satchels, rockets, etc. in a more
realistic spot. For example, in boot_camp you can place satchels over the
lamp that is underneath the gauss (hornet/LJ location, not bridge), and
without the fix those satchels simply explode on the floor above, outside
the building. You can sit at the staircase next to the lamp with 10 satchels
on the lamp, explode them and you don't receive any damage; only someone
picking up the gauss will receive damage, because the explosion actually
occurs up there, at the other side of the ceiling. Something similar happens
to nades, rockets, etc. The explosion spot is not really where you would
think it is (where you shoot at), but some 45-58 units in the opposite
direction of the projectile or in the case of satchels, above them. This
cvar fixes that, meaning lots of nukes related to explosive weapons also
get fixed with this (I don't know if all of them), and fixes the boot_camp
case of satchels on the lamp, which the `sv_ag_nuke_satchel 0` cvars cannot
fix because it's simply not a nuke, it's this other bug of explosion
positioning. It also makes nade-jumping a bit less risky, because the nade
explodes closer to the floor so you don't get as much damage, and by the
same logic it also nerfs throwing nades at opponents' feet, as they receive
a bit less damage, the explosion happening a bit further from their model
center/origin. Not completely related but it also fixes direct hits from
explosives, as they would sometimes deal 99hp damage instead of 100hp in
the case of nades, so a direct hit now deals the damage specified by the
corresponding `sv_ag_dmg_*` cvar. And throwing a nade/rocket at someone's
face when really close no longer should result in these weird cases of
you suiciding and the other one not getting hurt at all, or the opposite,
you not receiving any damage and the other dying. Note that this doesn't
fix using someone as meatshield to avoid damage. If you throw the nade at
the wall behind a player and you quickly position yourself between the nade
and the player, you will still avoid damage and the other player will
probably die, like when you throw a nade out of a door and quickly take
cover behind the wall before the nade lands and explodes.

In addition to these changes, `ag_gauss_fix 2` is now separated into
2 and 3. So it's now like this:
* ag_gauss_fix 0 - All selfgauss enabled
* ag_gauss_fix 1 - No self damage from your own gauss reflections when they
break through the wall (wallgauss) and bounce directly back to you
* ag_gauss_fix 2 - No self damage from your own gauss nukes (lamps, etc.)
* ag_gauss_fix 3 - No self damage from any of your own gauss shots, including
reflecting the gauss on the floor or wall in an angle that doesn't wallgauss