Add dirty state indicator to game name
The game name now will have an asterisk appended if some gamemode-related
cvar has been modified. It will show an extra asterisk every 5 changed
cvars, to indicate approximately how impure the gamemode is at that moment.

So for example:
0 cvars changed  -> AG TDM
1 cvar changed   -> AG TDM*
4 cvars changed  -> AG TDM*
5 cvars changed  -> AG TDM**
10 cvars changed -> AG TDM***

And so on... if those cvars are defaulted again, either individually or
with a successful `agdefault` vote, the game name will be updated
accordingly. So if it had 10 changed cvars (AG TDM***) and 1 of them is
defaulted, it will be back to 2 asterisks (AG TDM**)