Add a speedrun mode for fraglimited games
New cvar:
* sv_ag_speedrun <0|1> - Default: 0 (don't tell time at the end)

When set to 1 it simply tells the time it took to finish the game if
`mp_fraglimit` is not 0. It can be either just a normal game without
agstart or a match with agstart. If it's a match then it will start
counting the time when the match starts (when the countdown reaches 0
and players spawn), otherwise it will start when the map loads