Make team change penalty customizable
New cvar:
* sv_ag_match_teamchange_suicide_penalty <number> - Default: 1

If set to 0, there's no penalty when changing team during a match. If set
to 1, it removes a frag and adds a death. If set to 5, it removes 5 frags
and adds 5 deaths, and so on...

Sometimes 1 frag penalty is not enough, so it's now votable and each
gamemode can set the penalty they wish