Add spawn seeding + some cleanup
New cvar:
* ag_spawn_seed <0|number> - Default: 0 (use classic engine seeding)

This allows to have predictable spawn sequences. So if you set the cvar to
50 and do `agstart`, you will see a sequence, and if you do the same in
another server, with the same seed, you will see the same sequence of spawns.
If you change the seed, the sequence will change, but it will be the same
sequence in any AG 6.7+ server as long as you have the same seed.

The PRNG in the engine has an epoch-based seed and uses a LCG-like algorithm
to calculate the next number, and it's also used everywhere in the game, so
shooting the shotgun for example would make the next spawn change, because
the shotgun spread uses the same PRNG as the spawn selection, and that
generated number would be used for the next spawn instead if it hadn't been
used for the shotgun.

So the default behaviour would be to continue using the engine PRNG, and
that's when `ag_spawn_seed` is 0. For any other value, a custom PRNG will
be used, and that value will be passed to it as the seed. So the seed won't
change automatically like the engine one, and it will be used only for the
spawn selection, thus ensuring that we keep the spawn sequence predictable.
This is useful for speedrunning and for practicing. I wouldn't recommend
to use custom seeding for normal playing sessions with other players, unless
everyone agrees on having predictable respawns.

The custom PRNG works for every spawn system, but it's not very predictable
in `ag_spawn_system 2` and `3`, due to their nature, of depending on distances
to enemies and other stuff. And also take into account that, in the classic
spawn system (`ag_spawn_system 0`), if you block a spawn that was going to be
selected, it will choose the next one instead, and the rest of the sequence
will be shifted +1 too.

The reseeding occurs only at map start and match start, meaning that if you
set a new seed, you have to either `agstart` or `agmap`/`changelevel` for it
to take effect. It's possible to make it so that it updates in a short time,
using the cvar tracking system implemented a few days ago, but it's not a
thing to be prioritized at the moment.