Add cvar logging options
New cvar:
* sv_ag_log_cvars <0-3> - Default: 0 (don't log any extra cvar stuff)

Making most of the cvars have the FCVAR_SERVER flag so that we can know
when they are changed results in a lot of clutter in the log, when the
`log on` command is set and `mp_logfile 1` or `mp_logecho 1`, which is
common for servers. HLDS would spit a huge lot of cvar lines at the
beginning of every map, both about cvars that exist and their values and
about cvar value changes, which happen a lot with this cvar tracking
system, because the way it works is it sets a bogus value and then detect
if it's different upon config execution, so a lot of cvars are set and they
are all output to console/log by the engine. This new cvar tries to limit
the amount of clutter you get and display just enough relevant cvar info,
like what cvars are changed at the beginning/end of a map, which is useful
to quickly see if a match didn't comply with the settings.

So the cvar can take 4 values:
* On 0 it will log a few cvar states at the beginning of a map, it's more
or less like in AG 6.6 but a few less cvars.
* On 1 it will also log the changed cvars at the start/end of a map.
* On 2 it will also log all the gamemode cvars with their current values
at the start of a map. They're quite a lot, this is clutter but much less
than what it was before this commit.
* On 3 it will log every cvar value change made by the cvar system. Meant
for debugging purposes. It's verbose.