Add a way to move freely during changelevel delay
New command:
* delay_changelevel

This is for singleplayer. When you execute `delay_changelevel`, it prints
to console whether changelevel delaying is enabled or disabled, so that
you can know its state, only on `developer 1` or more. When it's enabled,
you can go through or use `trigger_changelevel` entities without doing
the actual map change, and when you execute the command again, it will do
the map change. So you can bind this to some key and press that key before
a changelevel, like when you are destroying the boxes at the end of Blast
Pit, and then you can go through the changelevel trigger, bhop back to the
boxes and press that key again to change to the next map, and you would
appear after the barricade of PU in that example. In the CLD of the Hazard
Course you can bhop and do whatever you want from the elevator doors until
the place where you want to change the map, instead of having a long script
with annoying movement where any minor mistake throws the run