Add unlimited uranium mode for Gauss% runs or gamemodes
New cvar:
* sv_ag_unlimited_uranium <0|1> - Default: 0 (normal)

Set to 1 so that gauss and egon don't use any uranium.

Useful for a Gauss% run of HL1 where you only use gauss. And for a new
version of Instagib and a new egon-only gamemode for tracking aim
practice.

Any uranium-consuming weapons added by plugins should account for this
cvar to decide whether to spend uranium.